refactor: reorganize type definitions into separate files for better structure
