+++ src/common/types/color-theme-kind.type.ts
@@ -1,0 +1,5 @@
+export const enum ColorThemeKind {
+  light = 1,
+  dark = 2,
+  highContrast = 3,
+}
