+++ src/common/types/ext-position.type.ts
@@ -1,0 +1,4 @@
+export const enum ExtPosition {
+  panel = 'panel',
+  explorer = 'explorer',
+}
