+++ src/common/types/index.ts
@@ -1,0 +1,5 @@
+export * from './pokemon.type';
+export * from './ext-position.type';
+export * from './theme.type';
+export * from './color-theme-kind.type';
+export * from './webview-message.type';
