+++ src/common/types/theme.type.ts
@@ -1,0 +1,6 @@
+export const enum Theme {
+  none = 'none',
+  forest = 'forest',
+  castle = 'castle',
+  beach = 'beach',
+}
