+++ src/common/types/webview-message.type.ts
@@ -1,0 +1,9 @@
+export class WebviewMessage {
+  text: string;
+  command: string;
+
+  constructor(text: string, command: string) {
+    this.text = text;
+    this.command = command;
+  }
+}
